UI面メイン

+++ Assets/Scripts/main/Unit/Effect.cs
@@ -1,9 +1,9 @@
 public class Effect
 {
-    public float hp_buff { get; protected set; } = 0;
-    public float atk_buff { get; protected set; } = 0;
-    public float def_buff { get; protected set; } = 0;
-    public float spd_buff { get; protected set; } = 0;
+    public int hp_buff { get; protected set; } = 0;
+    public int atk_buff { get; protected set; } = 0;
+    public int def_buff { get; protected set; } = 0;
+    public int spd_buff { get; protected set; } = 0;
     public int move_range_buff { get; protected set; } = 0;
     public virtual void Init(Unit owner) {}
     public virtual void Disable(Unit owner) {}
